Improved version almost ready for experiimentation with humans. Introduced a minimal baseline, which is going to be used in the experiments and essentially is just a chatbot with no other particular features. Improved the general vocabulary of the robot. Improved the social qualification and the associated rules (now if nao has a subordinate role, this will affect not only suffix and prefix of the sentence but also the way the sentence is said, i.e., faster and with higher voice, based on fuzzy inference). Done some small refactory. Introduced also support to additional animations, and a goodbye social practice where the robot goes to sleep. This commit also contains some changes in the code for the analysis of the results from the simulations about learning.

+++ sar/utils/weather.py
@@ -21,7 +21,7 @@ def getCurrentWeather():
             # store the value of "main"
             # key in variable y
             y = x["main"]
-            current_temperature = round(float(y["temp"])-273.15,0)
+            current_temperature = int(round(float(y["temp"])-273.15,0))
             # current_pressure = y["pressure"]
             current_humidity = float(y["humidity"])
             if current_humidity > 60:
